move email regex impl to it own package, rename to be more specific
