+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/validation/validateEmail/validateEmailImpls/ValidateEmailAndroidImpl.kt
@@ -1,8 +1,10 @@
-package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail
+package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.validateEmailImpls
 
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.IEmailMatcher
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.IValidateEmail
 import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherAndroidImpl
 
-class ValidateEmailImpl (
+class ValidateEmailAndroidImpl (
     private val emailMatcher: IEmailMatcher = EmailMatcherAndroidImpl()
 ): IValidateEmail {
     override fun validate(email: String): Boolean {
