+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/validation/validateEmail/validateEmailImpls/ValidateEmailRegexImpl.kt
@@ -1,5 +1,7 @@
-package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail
+package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.validateEmailImpls
 
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.IEmailMatcher
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.IValidateEmail
 import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherRegexImpl
 
 class ValidateEmailRegexImpl (
