+++ app/src/main/java/com/realityexpander/tasky/di/AppModule.kt
@@ -15,7 +15,7 @@ import com.realityexpander.tasky.auth_feature.domain.IAuthRepository
 import com.realityexpander.tasky.auth_feature.domain.validation.ValidatePassword
 import com.realityexpander.tasky.auth_feature.domain.validation.ValidateUsername
 import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.IValidateEmail
-import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.ValidateEmailImpl
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.validateEmailImpls.ValidateEmailAndroidImpl
 import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherAndroidImpl
 import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherRegexImpl
 import dagger.Module
@@ -108,13 +108,13 @@ object AppModule {
     @Singleton
     @Named("ValidateEmailAndroid")
     fun provideValidateEmailAndroid(): IValidateEmail =
-        ValidateEmailImpl(emailMatcher = EmailMatcherAndroidImpl())
+        ValidateEmailAndroidImpl(emailMatcher = EmailMatcherAndroidImpl())
 
     @Provides
     @Singleton
     @Named("ValidateEmailRegex")
     fun provideValidateEmailRegex(): IValidateEmail =
-        ValidateEmailImpl(emailMatcher = EmailMatcherRegexImpl())
+        ValidateEmailAndroidImpl(emailMatcher = EmailMatcherRegexImpl())
 
     @Provides
     @Singleton
